scaffolder-template-backstage/scaffolder-template-software/scaffolder-python-fastapi/uday-app/app/services/auth.py
from datetime import datetime, timedelta
from typing import Optional
from passlib.context import CryptContext
from jose import JWTError, jwt
from fastapi import Depends, HTTPException, status
from fastapi.security import OAuth2PasswordBearer
from sqlalchemy.orm import Session
from app.core.config import settings
from app.core.database import get_db
from app.models.user import User
from app.schemas.user import UserCreate, TokenData

pwd_context = CryptContext(
    schemes=["bcrypt"],
    deprecated="auto",
    bcrypt__default_rounds=12,
    bcrypt__default_ident="2b"
)
oauth2_scheme = OAuth2PasswordBearer(tokenUrl="auth/login")

def verify_password(plain_password: str, hashed_password: str) -> bool:
    return pwd_context.verify(plain_password, hashed_password)

def get_password_hash(password: str) -> str:
    # Garantir que a senha não exceda 72 bytes (limitação do bcrypt)
    password = password[:72]
    return pwd_context.hash(password)

def authenticate_user(db: Session, email: str, password: str) -> Optional[User]:
    user = db.query(User).filter(User.email == email).first()
    # Truncar a senha antes de verificar
    password = password[:72]
    if not user or not verify_password(password, user.senha_hash):
        return None
    return user

def create_access_token(data: dict, expires_delta: Optional[timedelta] = None) -> str:
    to_encode = data.copy()
    if expires_delta:
        expire = datetime.utcnow() + expires_delta
    else:
        expire = datetime.utcnow() + timedelta(minutes=15)
    to_encode.update({"exp": expire})
    encoded_jwt = jwt.encode(to_encode, settings.SECRET_KEY, algorithm=settings.JWT_ALGORITHM)
    return encoded_jwt

async def get_current_user(
    token: str = Depends(oauth2_scheme),
    db: Session = Depends(get_db)
) -> User:
    credentials_exception = HTTPException(
        status_code=status.HTTP_401_UNAUTHORIZED,
        detail="Could not validate credentials",
        headers={"WWW-Authenticate": "Bearer"},
    )
    try:
        payload = jwt.decode(token, settings.SECRET_KEY, algorithms=[settings.JWT_ALGORITHM])
        email: str = payload.get("sub")
        if email is None:
            raise credentials_exception
        token_data = TokenData(email=email)
    except JWTError:
        raise credentials_exception
    
    user = db.query(User).filter(User.email == token_data.email).first()
    if user is None:
        raise credentials_exception
    return user

async def create_user(user: UserCreate, db: Session) -> Optional[User]:
    # Verificar se o usuário já existe
    db_user = db.query(User).filter(User.email == user.email).first()
    if db_user:
        return None
    
    # Criar novo usuário
    hashed_password = get_password_hash(user.senha)
    # Definir avatar padrão baseado no gênero
    default_avatar = "homem.png" if user.genero == "homem" else "mulher.png"
    db_user = User(
        email=user.email,
        senha_hash=hashed_password,
        nome=user.nome,
        bio=user.bio,
        profissao=user.profissao,
        departamento=user.departamento,
        genero=user.genero,
        foto_perfil=default_avatar
    )
    db.add(db_user)
    db.commit()
    db.refresh(db_user)
    return db_user