scaffolder-template-backstage/scaffolder-template-software/scaffolder-python-fastapi/uday-app/requirements.txt
fastapi==0.104.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic[email]==2.5.1
pydantic-settings==2.1.0
typing-extensions==4.8.0
alembic==1.12.1
pytest==7.4.3
python-magic==0.4.27
uvicorn==0.24.0
